% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{phi_pairs}
\alias{phi_pairs}
\title{Calculate proportions of group k relative to reference group r}
\usage{
phi_pairs(n_events)
}
\arguments{
\item{n_events}{Event counts in group k and r}
}
\value{
A vector that has length of \eqn{C^k_{2}}
}
\description{
Calculate proportions of group k relative to reference group r
}
\examples{
phi_pairs(c(2557, 186, 176, 467, 120))
}
