% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmsamplesize.r
\name{pmsamplesize}
\alias{pmsamplesize}
\title{Sample Criteria Extension for Multinomial Regression}
\usage{
pmsamplesize(
  k,
  Q = NULL,
  props,
  n_events,
  shrinkage = 0.9,
  r2_cs = NULL,
  r2_nagelkerke = 0.15,
  simulation = c("glm", "lrm"),
  sim_params = NULL,
  sigma = 0.05
)
}
\arguments{
\item{k}{An integer. The number of levels in response variable}

\item{Q}{An integer. The number of candidate predictors. See Details}

\item{props}{A numeric vector. The proportions of each events}

\item{n_events}{A integer vector. The number of each events}

\item{shrinkage}{Default is 0.9}

\item{r2_cs}{A vector}

\item{sim_params}{A list of pairwise C statistics and prevalences}

\item{sigma}{}
}
\description{
Sample Criteria Extension for Multinomial Regression
}
