% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{props_by_events}
\alias{props_by_events}
\title{Calculate (cohort) outcome proportions}
\usage{
props_by_events(x)
}
\arguments{
\item{x}{The count of events of different outcomes}
}
\value{
A numeric vector of length \eqn{k + C^{k}_{2}} where the first k elements are
\eqn{p_k} and the rest are \eqn{p_{k,r}}
}
\description{
Calculate (cohort) outcome proportions
}
\examples{
props_by_events(c(2557, 186, 176, 467, 120))
}
