% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.criteria-package.R
\docType{package}
\name{sample.criteria-package}
\alias{sample.criteria}
\alias{sample.criteria-package}
\title{sample.criteria: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://chuckleong21.github.io/sample-criteria/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
