% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmsamplesize.r
\name{pmsamplesize}
\alias{pmsamplesize}
\title{Sample Criteria Extension for Multinomial Regression}
\usage{
pmsamplesize(
  k,
  Q = NULL,
  props,
  n_events,
  shrinkage = 0.9,
  r2_cs = NULL,
  r2_nagelkerke = 0.15,
  simulation = c("glm", "lrm"),
  sim_params = NULL,
  sigma = 0.05
)
}
\arguments{
\item{k}{An integer. The number of levels in response variable}

\item{Q}{An integer. The number of candidate predictors. See Details}

\item{props}{A numeric vector. The proportions of each events}

\item{n_events}{A integer vector. The number of each events}

\item{shrinkage}{A pre-defined shrinkage factor. Default to 0.9}

\item{r2_cs}{A numeric vector. Pre-defined Cox-Snell \eqn{R^2}.}

\item{r2_nagelkerke}{An adjustment for Cox-Snell \eqn{R^2}. Default to 0.15.}

\item{simulation}{Either \code{glm} or \code{lrm}. See details.}

\item{sim_params}{A list of pairwise C statistics and prevalences}

\item{sigma}{A single double. tolerate rate between adjusted \eqn{R^2} and apparent \eqn{R^2}}
}
\value{
An integer
}
\description{
Calculate sample criteria basing on multinominal regression. The criteria are
defined in Riley's work.
}
\details{
Additional details...
}
\examples{
K <- 5
events <- c(2557, 186, 176, 467, 120)
c_stats <- c(0.85, 0.92, 0.99, 0.95, 0.75, 0.95, 0.87, 0.87, 0.71, 0.82)
prev <- phi_pairs(events)
set.seed(101)
pmsamplesize(k = K, n_events = events,
simulation = "glm", sim_params = list(c_stats = c_stats, prev = prev))
}
\references{
\insertCite{sample.criteria-package}{sample.criteria}
}
