% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmsamplesize.r
\name{pmsamplesize}
\alias{pmsamplesize}
\title{Sample Criteria Extension for Multinomial Regression}
\usage{
pmsamplesize(
  Q,
  k,
  p,
  r2_cs = NULL,
  auc = NULL,
  prev,
  shrinkage = 0.9,
  r2_nagelkerke = 0.15,
  sigma = 0.05,
  margin_error = 0.05
)
}
\arguments{
\item{Q}{An integer. The number of candidate predictors. See Details}

\item{k}{An integer. The number of levels in response variable}

\item{p}{A numeric vector. Outcome proportion(s)}

\item{r2_cs}{numeric. Cox-Snell \eqn{R^2}.}

\item{auc}{numeric. Pairwise C statistics}

\item{prev}{numeric. Outcome proportions in categories relative to reference}

\item{shrinkage}{A pre-defined shrinkage factor. Default to 0.9}

\item{r2_nagelkerke}{An adjustment for Cox-Snell \eqn{R^2}. Default to 0.15.}

\item{sigma}{A double. tolerate rate between adjusted \eqn{R^2} and apparent \eqn{R^2}. Default to 0.05}

\item{margin_error}{A divergence for margin error specified by Chi-sqaure distribution's confidence level.
Deault to 0.05}
}
\value{
An integer
}
\description{
Calculate sample criteria extended for multinominal regression.
}
\details{
Sample Criteria for Clinical Prediction Model are based on the work of
\href{https://onlinelibrary.wiley.com/doi/full/10.1002/sim.7992}{Riley 2018} and
\href{https://onlinelibrary.wiley.com/doi/epdf/10.1002/sim.8806}{Calster 2020}
}
\examples{
K <- 5
events <- c(2557, 186, 176, 467, 120)
c_stats <- c(0.85, 0.92, 0.99, 0.95, 0.75, 0.95, 0.87, 0.87, 0.71, 0.82)
prev <- phi_pairs(events)
set.seed(1234)
pmsamplesize(Q = 30, k = 2, auc = 0.81, prev = 0.77)
}
\references{
\insertCite{pmsamplesize}{sample.criteria}
}
