% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmsamplesize.r
\name{pmsamplesize}
\alias{pmsamplesize}
\title{Sample Criteria Extension for Multinomial Regression}
\usage{
pmsamplesize(
  Q,
  k,
  p,
  adjust = FALSE,
  r2_cs_app = NULL,
  r2_cs_adj = NULL,
  shrinkage = 0.9,
  r2_nagelkerke = 0.15,
  auc = NULL,
  prev = NULL,
  sigma = 0.05,
  margin_error = 0.05
)
}
\arguments{
\item{Q}{An integer. The number of candidate predictors. See Details}

\item{k}{An integer. The number of levels in response variable}

\item{p}{A numeric vector. Outcome proportion(s) or event observations}

\item{adjust}{A logical. Should adjusted R-square value be used? Default to FALSE}

\item{r2_cs_app}{A numeric vector. The apparent Cox-Snell \eqn{R^2}}

\item{r2_cs_adj}{A numeric vector. The optimism-adjusted Cox-Snell \eqn{R^2}}

\item{shrinkage}{A pre-defined shrinkage factor. Default to 0.9}

\item{r2_nagelkerke}{An adjustment for Cox-Snell \eqn{R^2}. Default to 0.15.}

\item{auc}{A numeric vector. Pairwise C statistics}

\item{prev}{A numeric vector. Outcome proportions in categories relative to reference}

\item{sigma}{A double. Tolerate rate between adjusted \eqn{R^2} and apparent \eqn{R^2}. Default to 0.05}

\item{margin_error}{A double. A divergence for margin error specified by
Chi-square distribution's confidence level. Default to 0.05}
}
\value{
A \code{pmsample} class object
\itemize{
\item \strong{final}: A single row data frame representing the selected criterion for
the minimum sample size for developing a new model
\item \strong{criteria}: A data frame containing all the sample sizes
based on three criteria. Divergences in sample sizes due to number rounding
\item sigma: A single numeric representing the tolerate rate between
adjusted \eqn{R^2} and apparent \eqn{R^2}
\item margin_error: A divergence for margin error specified by
Chi-square distribution's confidence level
}
}
\description{
Calculate sample criteria extended for multinominal regression.
}
\details{
The pmsample objects created by \code{pmsamplesize} can represent the minimum
sample size needed for developing a new model based on the current studies.

\code{pmsamplesize} function adapts the amount of information based on
the type of model. A certain parameters need pre-specifying for calculation:
\itemize{
\item The number of candidate predicators \emph{Q}, usually obtained using
the fractional polynomial method and it is not addressed by this package.
\item The number of levels in response variable \emph{k}, typically for
a binary model k = 2, for multinomial models k > 2. One of the argument
between \code{k} or \code{p}, the outcome proportions or the observation counts
of the events, must be specified. In the case of a missing \code{k}, it will
then defined by the length of \code{p}.
\item The order of p matters for calculation: for a binary model, the second value
in the vector is perceived as 'outcome proportion' or 'outcome events'.
\item Pairwise C-statistics \code{auc} for simulation. The process requires
estimates of the pairwise outcome proportions \eqn{\phi_{k,r}} of
category \emph{k} relative to the reference category \emph{r}, \code{prev}.
If not supplied, they will be derived from \code{p}.
}
}
\examples{
# when only C statistics is reported in a binary model
set.seed(1234)
pmsamplesize(Q = 30, k = 2, r2_nagelkerke = 0.15, auc = 0.81, prev = 0.77)
}
\references{
\insertCite{Pate2023-yh}{sample.criteria}
\insertCite{Riley2019-rn}{sample.criteria}
}
