% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.criteria-package.R
\docType{package}
\name{sample.criteria-package}
\alias{sample.criteria}
\alias{sample.criteria-package}
\title{sample.criteria: Criteria-based minimum sample size for multinomial logistic models}
\description{
The package facilitate to derive minimum sample size for both binary and multinomial logistic models in clinic studies based on three criteria. 1) fixed shrinkage rate based 2) adjusted shrinkage rate by tolerate rate in models based and 3) Overall risk restrainted by margin error of distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jingwen517/sample-criteria.git}
  \item \url{https://chuckleong21.github.io/sample-criteria/}
}

}
\author{
\strong{Maintainer}: Jingwen Sun \email{jingwen7841@gmail.com}

}
\keyword{internal}
